Detect for ARM64 also

+++ winget-install.ps1
@@ -187,56 +187,87 @@ function Confirm-Exist ($AppID) {
 
 #Check if modifications exist in "mods" directory
 function Test-ModsInstall ($AppID) {
+
+    #Takes care of a null situation
+    $ModsPreInstall = $null
+    $ModsInstallOnce = $null
+    $ModsInstall = $null
+    $ModsUpgrade = $null
+
+    if (Test-Path "$PSScriptRoot\mods\$AppID-preinstall.ps1") {
+        $ModsPreInstall = "$PSScriptRoot\mods\$AppID-preinstall.ps1"
+    }
+
     if (Test-Path "$PSScriptRoot\mods\$AppID-install-once.ps1") {
         $ModsInstallOnce = "$PSScriptRoot\mods\$AppID-install-once.ps1"
-        return $ModsInstallOnce
+        return $ModsPreInstall, $ModsInstallOnce
     }
     elseif (Test-Path "$PSScriptRoot\mods\$AppID-install.ps1") {
         $ModsInstall = "$PSScriptRoot\mods\$AppID-install.ps1"
-        return $ModsInstall
+        return $ModsPreInstall, $ModsInstall
     }
     elseif (Test-Path "$PSScriptRoot\mods\$AppID-upgrade.ps1") {
         $ModsUpgrade = "$PSScriptRoot\mods\$AppID-upgrade.ps1"
-        return $ModsUpgrade
+        return $ModsPreInstall, $ModsUpgrade
     }
-    else {
-        return 0
-    }
+    return $ModsPreInstall
 }
 
 function Test-ModsUninstall ($AppID) {
     if (Test-Path "$PSScriptRoot\mods\$AppID-uninstall.ps1") {
         $ModsUninstall = "$PSScriptRoot\mods\$AppID-uninstall.ps1"
-        return $ModsUninstall
     }
-    else {
-        return 0
+    if (Test-Path "$PSScriptRoot\mods\$AppID-uninstalled.ps1") {
+        $ModsUninstalled = "$PSScriptRoot\mods\$AppID-uninstalled.ps1"
     }
+    return $ModsUninstall, $ModsUninstalled
 }
 
 #Install function
 function Install-App ($AppID, $AppArgs) {
     $IsInstalled = Confirm-Install $AppID
     if (!($IsInstalled)) {
+        #Check if mods exist for preinstall/install/upgrade
+        $ModsPreInstall, $ModsInstall = Test-ModsInstall $($AppID)
+
+        #Check if an preinstall mod already exist
+        if (!($ModsPreInstall) -and (Test-Path "$WAUInstallLocation\mods\$AppID-preinstall.ps1")) {
+            $ModsPreInstall = "$WAUInstallLocation\mods\$AppID-preinstall.ps1"
+        }
+
+        #If PreInstall script exist
+        if ($ModsPreInstall) {
+            Write-Log "-> Modifications for $AppID before upgrade are being applied..." "Yellow"
+            & "$ModsPreInstall"
+        }
+
         #Install App
         Write-Log "-> Installing $AppID..." "Yellow"
         $WingetArgs = "install --id $AppID -e --accept-package-agreements --accept-source-agreements -h $AppArgs" -split " "
         Write-Log "-> Running: `"$Winget`" $WingetArgs"
+
         & "$Winget" $WingetArgs | Tee-Object -file $LogFile -Append
 
         #Check if install is ok
         $IsInstalled = Confirm-Install $AppID
         if ($IsInstalled) {
             Write-Log "-> $AppID successfully installed." "Green"
-            #Check if mods exist
-            $ModsInstall = Test-ModsInstall $AppID
-            if ($ModsInstall -like "*$AppID-install*") {
-                Write-Log "-> Modifications for $AppID after install are being applied..." "Yellow"
-                & "$ModsInstall"
-            }
-            #Add to WAU mods if exists
+            #Check if an install/upgrade mod exist
             if (($ModsInstall -like "*$AppID-install*") -or ($ModsInstall -like "*$AppID-upgrade*")) {
+                if ($ModsInstall -like "*$AppID-install*") {
+                    Write-Log "-> Modifications for $AppID after install are being applied..." "Yellow"
+                    & "$ModsInstall"
+                }
+                #Add mods if deployed from app install
                 Add-WAUMods $AppID
+            }
+            else {
+                #Check if an install mod already exist
+                $ModsInstall = "$WAUInstallLocation\mods\$AppID-install.ps1"
+                if (Test-Path "$ModsInstall") {
+                    Write-Log "-> Modifications for $AppID after install are being applied..." "Yellow"
+                    & "$ModsInstall"
+                }
             }
             #Add to WAU White List if set
             if ($WAUWhiteList) {
@@ -263,17 +294,44 @@ function Uninstall-App ($AppID, $AppArgs) {
         & "$Winget" $WingetArgs | Tee-Object -file $LogFile -Append
 
         #Check if mods exist
-        $ModsUninstall = Test-ModsUninstall $AppID
+        $ModsUninstall, $ModsUninstalled = Test-ModsUninstall $AppID
         if ($ModsUninstall) {
             Write-Log "-> Modifications for $AppID during uninstall are being applied..." "Yellow"
             & "$ModsUninstall"
         }
+        else {
+            #Check if an uninstall mod already exist
+            $ModsUninstall = "$WAUInstallLocation\mods\$AppID-uninstall.ps1"
+            if (Test-Path "$ModsUninstall") {
+                Write-Log "-> Modifications for $AppID during uninstall are being applied..." "Yellow"
+                & "$ModsUninstall"
+            }
+        }
+        
         #Check if uninstall is ok
         $IsInstalled = Confirm-Install $AppID
         if (!($IsInstalled)) {
             Write-Log "-> $AppID successfully uninstalled." "Green"
-            #Remove from WAU mods (always)
-            Remove-WAUMods $AppID
+            if ($ModsUninstalled) {
+                Write-Log "-> Modifications for $AppID after uninstall are being applied..." "Yellow"
+                & "$ModsUninstalled"
+                #Remove mods if deployed from app install
+                if ((Test-Path "$PSScriptRoot\mods\$AppID-install.ps1") -or (Test-Path "$PSScriptRoot\mods\$AppID-upgrade.ps1")) {
+                    Remove-WAUMods $AppID
+                }
+            }
+            else {
+                #Check if an uninstalled mod already exist
+                $ModsUninstalled = "$WAUInstallLocation\mods\$AppID-uninstalled.ps1"
+                if (Test-Path "$ModsUninstalled") {
+                    Write-Log "-> Modifications for $AppID after uninstall are being applied..." "Yellow"
+                    & "$ModsUninstalled"
+                }
+                #Remove mods if deployed from app install
+                if ((Test-Path "$PSScriptRoot\mods\$AppID-install.ps1") -or (Test-Path "$PSScriptRoot\mods\$AppID-upgrade.ps1")) {
+                    Remove-WAUMods $AppID
+                }
+            }
             #Remove from WAU White List if set
             if ($WAUWhiteList) {
                 Remove-WAUWhiteList $AppID
@@ -320,7 +378,7 @@ function Add-WAUMods ($AppID) {
     $Mods = "$WAUInstallLocation\mods"
     if (Test-Path $Mods) {
         #Add mods
-        if ((Test-Path "$PSScriptRoot\mods\$AppID-install.ps1") -or (Test-Path "$PSScriptRoot\mods\$AppID-upgrade.ps1")) {
+        if ((Test-Path "$PSScriptRoot\mods\$AppID-preinstall.ps1") -or (Test-Path "$PSScriptRoot\mods\$AppID-install.ps1") -or (Test-Path "$PSScriptRoot\mods\$AppID-upgrade.ps1")) {
             Write-Log "-> Add modifications for $AppID to WAU 'mods'"
             Copy-Item "$PSScriptRoot\mods\$AppID-*" -Destination "$Mods" -Exclude "*-install-once*", "*-uninstall*" -Force
         }
@@ -334,7 +392,7 @@ function Remove-WAUMods ($AppID) {
     if (Test-Path "$Mods\$AppID*") {
         Write-Log "-> Remove $AppID modifications from WAU 'mods'"
         #Remove mods
-        Remove-Item -Path "$Mods\$AppID*" -Force
+        Remove-Item -Path "$Mods\$AppID*" -Exclude "*-uninstall*"-Force
     }
 }
